--- lib/data/index.ts
@@ -1,3 +1,0 @@
-export * from './get';
-export * from './types';
-export * from './yasb2xws';
