--- lib/data/types.ts
@@ -1,17 +1,0 @@
-export interface Ship {
-  id: string;
-  name: string;
-  icon: string;
-  pilots: { [name: string]: Pilot };
-}
-
-export interface Pilot {
-  id: string;
-  name: string;
-  caption?: string;
-}
-
-export interface Upgrade {
-  id: string;
-  name: string;
-}
